--- a/Common/Brokerages/BybitBrokerageModel.cs
+++ b/Common/Brokerages/BybitBrokerageModel.cs
@@ -15,6 +15,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using QuantConnect.Benchmarks;
 using QuantConnect.Orders;
 using QuantConnect.Orders.Fees;