… .github/workflows/api-tests.yml
… .github/workflows/gh-actions.yml
… .github/workflows/rebase-org-branches.yml
… .github/workflows/regression-tests.yml
… .github/workflows/report-generator.yml
… .github/workflows/research-regression-tests.yml
… .github/workflows/virtual-environments.yml
… Optimizer.Launcher/QuantConnect.Optimizer.Launcher.csproj
… Optimizer/QuantConnect.Optimizer.csproj
… Queues/QuantConnect.Queues.csproj
… Report/QuantConnect.Report.csproj
… Research/QuantConnect.Research.csproj
… Tests/Indicators/PythonIndicatorNoinheritanceTests.cs
… Tests/Indicators/PythonIndicatorNoinheritanceTestsLegacy.cs
… Tests/Indicators/PythonIndicatorTests.cs
… Tests/Indicators/TestHelper.cs
… Tests/QuantConnect.Tests.csproj
… Tests/app.config
… ToolBox/QuantConnect.ToolBox.csproj